requirements.txt
pyserial==3.2.1
sha3==0.2.1
